+++ parallel.cc
@@ -14,7 +14,7 @@ typedef std::complex<double> Complex;
 #include <chrono>
 
 namespace utils {
-int ALIGNMENT=2*sizeof(Complex); // Must be a multiple of sizeof(Complex)
+size_t ALIGNMENT=2*sizeof(Complex); // Must be a multiple of sizeof(Complex)
 }
 
 using namespace utils;
