+++ parallel.h
@@ -14,7 +14,7 @@ extern size_t threshold;
 
 namespace parallel {
 extern size_t lastThreads;
-inline int get_thread_num()
+inline size_t get_thread_num()
 {
 #ifdef SINGLE_THREAD
   return 0;
@@ -23,11 +23,11 @@ inline int get_thread_num()
 #endif
 }
 
-inline int get_thread_num(unsigned int threads) {
+inline size_t get_thread_num(size_t threads) {
   return threads > 1 ? get_thread_num() : 0;
 }
 
-inline int get_max_threads()
+inline size_t get_max_threads()
 {
 #ifdef SINGLE_THREAD
   return 1;
